requirements.txt
requests
openai
elevenlabs
python-dotenv
sounddevice
soundfile
numpy